feat: simplified sol connect handler
