+++ src/components/wallet-modal/solana-wallet-connect.tsx
@@ -1,11 +1,9 @@
 'use client';
 
 import SolanaIcon from '@/components/icons/solana';
+import useSolConnectHandler from '@/services/wallets/use-sol-connect-handler';
 import { Network } from '@/types/web3';
-import { useWallet } from '@solana/wallet-adapter-react';
-import { useWalletModal } from '@solana/wallet-adapter-react-ui';
 
-import useConnectHandler from './use-connect-handler';
 import WalletModalButton from './wallet-modal-button';
 
 type Props = {
@@ -13,24 +11,14 @@ type Props = {
 };
 
 export default function SolanaWalletConnect({ onConnect }: Props) {
-  const { setVisible, visible } = useWalletModal();
-  const { publicKey, disconnect } = useWallet();
-
-  const address = publicKey?.toString();
-
-  useConnectHandler({
-    address,
-    visible,
-    onConnect: () => onConnect(address ?? '', Network.Sol),
-    disconnect,
-  });
+  const onOpenModal = useSolConnectHandler((address) =>
+    onConnect(address, Network.Sol)
+  );
 
   return (
     <WalletModalButton
       startIcon={<SolanaIcon sx={{ fontSize: '24' }} />}
-      onClick={() => {
-        setVisible(true);
-      }}
+      onClick={onOpenModal}
     >
       Solana
     </WalletModalButton>
